# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/dump/hotel/autocomplete": {
      "get": {
        "operationId": "HotelDumpController_getHotelAutocomplete",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": true,
            "in": "query",
            "description": "Search Query",
            "schema": {
              "example": "Cairo, Egypt, India, Mumbai, etc.",
              "type": "string"
            }
          },
          {
            "name": "lat",
            "required": false,
            "in": "query",
            "description": "User Latitude",
            "schema": {
              "example": "19.076",
              "type": "number"
            }
          },
          {
            "name": "long",
            "required": false,
            "in": "query",
            "description": "User Longitude",
            "schema": {
              "example": "72.8777",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get hotel autocomplete suggestions",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
    "/dump/hotel/details": {
      "get": {
        "operationId": "HotelDumpController_getHotelDetails",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotelId",
            "required": true,
            "in": "query",
            "description": "Hotel ID",
            "schema": {
              "example": "107953",
              "type": "string"
            }
          },
          {
            "name": "supplierCode",
            "required": true,
            "in": "query",
            "description": "Supplier Code",
            "schema": {
              "example": "Tbo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get hotel details by hotel code",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
    "/dump/hotel/room-content": {
      "get": {
        "operationId": "HotelDumpController_getHotelRoomContent",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hotelId",
            "required": true,
            "in": "query",
            "description": "Hotel ID",
            "schema": {
              "example": "107953",
              "type": "string"
            }
          },
          {
            "name": "supplierCode",
            "required": true,
            "in": "query",
            "description": "Supplier Code",
            "schema": {
              "example": "Tbo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Get hotel room content by hotel code",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
    "/dump/hotel/transfer-data": {
      "post": {
        "operationId": "HotelDumpController_transferDataToHotelContent",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferDataToHotelContent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Transfer data to hotel content table",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
    "/dump/hotel/country-list": {
      "post": {
        "operationId": "HotelDumpController_addCountryList",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Add country list dump from TBO API",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
    "/dump/hotel/city-list": {
      "post": {
        "operationId": "HotelDumpController_addCityList",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Add city list dump from TBO API",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
    "/dump/hotel/hotel-list": {
      "post": {
        "operationId": "HotelDumpController_addHotelList",
        "parameters": [
          {
            "name": "Currency-Preference",
            "in": "header",
            "description": "Please enter preferred currency code. (3 digit currency code)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "header",
            "description": "Please enter preferred language. (e.g., english,french)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Ip-Address",
            "in": "header",
            "description": "Please enter IP address. (e.g., 127.0.0.1)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Api-version",
            "in": "header",
            "description": "Please enter API version.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "v1"
              ]
            }
          },
          {
            "name": "Club-Id",
            "in": "header",
            "description": "Please enter club id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Device-Information",
            "in": "header",
            "description": "Please enter device-information id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Entity not found"
          },
          "422": {
            "description": "Unprocessable entity"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "Add hotel list dump from TBO API",
        "tags": [
          "Hotel Dump"
        ]
      }
    },
...
  },
  "info": {
    "title": "API SERVICE",
    "description": "## Travel API Documentation\n\nWelcome to the **API SERVICE** documentation. This guide provides comprehensive guidelines and detailed instructions on how to effectively integrate and utilize the Travel API.\n\n### Overview\nThis documentation includes various sections covering:\n- **Authentication Methods**: Learn how to authenticate using Bearer and Cookie authentication.\n- **Endpoint Descriptions**: Detailed information on each API endpoint.\n- **Usage Examples**: Practical examples to help you get started quickly.\n- **Best Practices**: Tips and recommendations for efficient implementation.\n\n### Purpose\nWhether you are a developer looking to incorporate travel-related functionalities or a business aiming to enhance your travel services, this documentation serves as a valuable resource to guide you through the process.\n\n### Getting Started\nTo get started, ensure you have the necessary authentication tokens and follow the endpoint descriptions to make your first API call.\n\nFor any questions or support, please refer to the contact section at the end of this documentation.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth"
      }
    },
    "schemas": {
...
      "TransferDataToHotelContent": {
        "type": "object",
        "properties": {
          "from": {
            "type": "number",
            "description": "Start Number",
            "example": 0,
            "minimum": 0
          },
          "to": {
            "type": "number",
            "description": "End Number",
            "example": 1000,
            "minimum": 0
          }
        },
        "required": [
          "from",
          "to"
        ]
      }
    }
  }
}